:recycle: Rename for adding reminder prefix
