+++ lib/domain/settings/setting_page.dart
@@ -7,7 +7,7 @@ import 'package:pilll/domain/settings/components/rows/creating_new_pillsheet.dar
 import 'package:pilll/domain/settings/components/rows/menstruation.dart';
 import 'package:pilll/domain/settings/components/rows/account_link.dart';
 import 'package:pilll/domain/settings/components/rows/list_explain.dart';
-import 'package:pilll/domain/settings/components/rows/notification_customize_word.dart';
+import 'package:pilll/domain/settings/components/rows/reminder_notification_customize_word.dart';
 import 'package:pilll/domain/settings/components/rows/notification_in_rest_duration.dart';
 import 'package:pilll/domain/settings/components/rows/notification_time.dart';
 import 'package:pilll/domain/settings/components/rows/pill_sheet_remove.dart';
@@ -154,7 +154,7 @@ class SettingPage extends HookConsumerWidget {
                         ),
                         _separator(),
                       ],
-                      NotificationCustomizeWord(
+                      ReminderNotificationCustomizeWord(
                           setting: setting,
                           isTrial: state.isTrial,
                           isPremium: state.isPremium,
